Found bug with fix case
Bug: we were not properly updating epsilon transition when replacing
(n' -> alpha ns) with (n' -> N ns) for each (n -> N).
If alpha has an epsilon transition, then ns in this case must be empty,
since we cannot have an epsilon preceding another production.
In this case then the types *should* line up. But I haven't looked into
explaining this to the typechecker, so I left an unsafeCoerce for now.